fixed proof testing
