+++ src/modules/credentials/handlers/ProposeCredentialHandler.ts
@@ -25,7 +25,7 @@ export class ProposeCredentialHandler implements Handler {
 
   public async handle(messageContext: HandlerInboundMessage<ProposeCredentialHandler>) {
     const credentialRecord = await this.credentialService.processProposal(messageContext)
-    if (await this.credentialAutoResponseCoordinator.shoudlAutoRespondToProposal(credentialRecord)) {
+    if (this.credentialAutoResponseCoordinator.shoudlAutoRespondToProposal(credentialRecord)) {
       return await this.createOffer(credentialRecord, messageContext)
     }
   }
